requirements.txt
amqp==2.5.2
APScheduler==3.6.3
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
dj-fine-search==0.0.4
Django==3.0.3
django-apscheduler==0.3.0
django-celery-beat==2.0.0
django-cors-headers==3.2.1
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
fpdf
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
kombu==4.6.8
lazy-object-proxy==1.4.3
mccabe==0.6.1
pandas
Pillow==7.0.0
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
python-crontab==2.4.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
twilio==6.37.0
typed-ast==1.4.1
tzlocal==2.0.0
urllib3==1.25.8
vine==1.3.0
wrapt==1.11.2
zipp==3.1.0
